+++ bridge.h
@@ -80,11 +80,11 @@ class Bridge : public KDecorationBridgeUnstable
         // Window tabbing
         virtual bool isClientGroupActive();
         virtual QList< ClientGroupItem > clientGroupItems() const;
-        virtual int itemId( int index );
+        virtual long itemId( int index );
         virtual int visibleClientGroupItem();
         virtual void setVisibleClientGroupItem( int index );
         virtual void moveItemInClientGroup( int index, int before );
-        virtual void moveItemToClientGroup( int itemId, int before );
+        virtual void moveItemToClientGroup( long itemId, int before );
         virtual void removeFromClientGroup( int index, const QRect& newGeom );
         virtual void closeClientGroupItem( int index );
         virtual void closeAllInClientGroup();
